requirements.txt
lief
python-magic
matplotlib
numpy==1.26.4
seaborn
secml==0.15.6
lightgbm
deap
git+https://github.com/zangobot/ember.git
requests
torch
torchvision